+++ marketing/src/containers/compliance/eudr-tool/constants.ts
@@ -1,41 +1,42 @@
-export const EUDRToolFeatures1 = [
-  {
-    index: '01',
-    title: 'Achieve EUDR Alignment and Beyond',
-    description:
-      'Align with new standards and requirements (ESRS, TNFD, SBTN). Produce transparent, repeatable, and auditable analysis.',
-  },
-  {
-    index: '02',
-    title: 'Produce Transparent, Repeatable, and Auditable Analyses',
-    description:
-      'Our tool ensures your analyses are transparent, repeatable, and audit-ready, instilling confidence in your sustainability reporting.',
-  },
-  {
-    index: '03',
-    title: 'Visualize and Analyze Risk Exposure',
-    description:
-      'Identify and understand risk exposure by commodity, volume, supplier, and region through intuitive visualizations and in-depth analysis.',
-  },
-];
-
-export const EUDRToolFeatures2 = [
-  {
-    index: '04',
-    title: 'Comprehensively Manage Scope 3 Supply Chain',
-    description:
-      'Our tool covers a wide range of commodities including cattle, cocoa, coffee, palm oil, rubber, soy, wood, and more, enabling comprehensive management of your scope 3 supply chain.',
-  },
-  {
-    index: '05',
-    title: 'Seamless Integration with Existing Data Systems',
-    description:
-      'Easily integrate our tool with your existing data systems, streamlining your sustainability efforts without disruption.',
-  },
-  {
-    index: '06',
-    title: 'Leverage Latest Scientific Data',
-    description:
-      'Make informed decisions backed by the latest scientific data, ensuring your sustainability strategies are grounded in sound knowledge and expertise.',
-  },
+export const EUDRToolFeatures = [
+  [
+    {
+      index: '01',
+      title: 'Achieve EUDR Alignment and Beyond',
+      description:
+        'Align with new standards and requirements (ESRS, TNFD, SBTN). Produce transparent, repeatable, and auditable analysis.',
+    },
+    {
+      index: '02',
+      title: 'Produce Transparent, Repeatable, and Auditable Analyses',
+      description:
+        'Our tool ensures your analyses are transparent, repeatable, and audit-ready, instilling confidence in your sustainability reporting.',
+    },
+    {
+      index: '03',
+      title: 'Visualize and Analyze Risk Exposure',
+      description:
+        'Identify and understand risk exposure by commodity, volume, supplier, and region through intuitive visualizations and in-depth analysis.',
+    },
+  ],
+  [
+    {
+      index: '04',
+      title: 'Comprehensively Manage Scope 3 Supply Chain',
+      description:
+        'Our tool covers a wide range of commodities including cattle, cocoa, coffee, palm oil, rubber, soy, wood, and more, enabling comprehensive management of your scope 3 supply chain.',
+    },
+    {
+      index: '05',
+      title: 'Seamless Integration with Existing Data Systems',
+      description:
+        'Easily integrate our tool with your existing data systems, streamlining your sustainability efforts without disruption.',
+    },
+    {
+      index: '06',
+      title: 'Leverage Latest Scientific Data',
+      description:
+        'Make informed decisions backed by the latest scientific data, ensuring your sustainability strategies are grounded in sound knowledge and expertise.',
+    },
+  ],
 ];
