+++ marketing/src/containers/compliance/hero/component.tsx
@@ -1,33 +1,30 @@
-import { PauseIcon, PlayIcon } from '@heroicons/react/solid';
+import { FC } from 'react';
+import { useEffect, useRef, useCallback } from 'react';
+import { useInView } from 'react-intersection-observer';
 import Wrapper from 'containers/wrapper';
-
+import { Media } from 'components/media-query';
+import Image from 'next/image';
+import YouTube from 'react-youtube';
 import { motion } from 'framer-motion';
-import { useCallback, useEffect, useRef, useState } from 'react';
 
-const Hero: React.FC = () => {
+const Video: FC = () => {
   const videoRef = useRef<HTMLVideoElement>();
-  const [playing, setPlaying] = useState(false);
-
-  const onTogglePlay = useCallback(() => {
-    if (videoRef.current) {
-      if (!playing) videoRef.current.play();
-      if (playing) videoRef.current.pause();
-
-      setPlaying(!playing);
-    }
-  }, [playing]);
+  const { ref, inView } = useInView();
 
   const onVideoEnded = useCallback(() => {
-    setPlaying(false);
     videoRef.current.pause();
     videoRef.current.currentTime = 0;
   }, []);
 
   useEffect(() => {
+    if (videoRef.current && !inView) {
+      videoRef.current.pause();
+    }
+
     if (videoRef.current) {
       videoRef.current.addEventListener('ended', onVideoEnded);
     }
-  }, [onVideoEnded]);
+  }, [inView, onVideoEnded]);
 
   return (
     <section className="relative z-10 bg-white">
@@ -44,41 +41,30 @@ const Hero: React.FC = () => {
             </h1>
           </div>
         </motion.div>
-
-        <motion.div
-          className="relative hidden aspect-video"
-          initial={{ opacity: 0, y: 20 }}
-          animate={{ opacity: 1, y: 0 }}
-          transition={{ duration: 0.75, delay: 0.1 }}
-        >
-          <header className="absolute z-10 flex items-center space-x-5 text-white top-6 left-6">
-            <button
-              className="flex items-center justify-center w-20 h-20 bg-white cursor-pointer"
-              onClick={onTogglePlay}
-            >
-              {!playing && <PlayIcon className="w-5 h-5 text-black" />}
-              {playing && <PauseIcon className="w-5 h-5 text-black" />}
-            </button>
-
-            <h4 className="text-sm font-black uppercase font-display">Watch demo</h4>
-          </header>
-          <motion.iframe
-            initial={{ opacity: 0, y: 20 }}
-            animate={{ opacity: 1, y: 0 }}
-            transition={{ duration: 0.75, delay: 0.1 }}
-            width="853"
-            height="480"
-            src="https://www.youtube.com/embed/RIJ1lX57TtI"
-            frameBorder="0"
-            allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-            allowFullScreen
-            title="Landgriffon: Sustainable Supply Chain Solutions"
-            className="w-full h-[190px] md:h-[438px] lg:h-[603px]"
-          />
-        </motion.div>
       </Wrapper>
+      <div className="relative z-20">
+        <Wrapper>
+          <section ref={ref} className="relative max-h-screen bg-white">
+            <YouTube
+              videoId="Bzvdwt-EuNM"
+              title="Landgriffon: Sustainable Supply Chain Solutions"
+              className="w-full h-[320px] md:h-[418px] lg:h-[583px]"
+              iframeClassName="w-full h-[320px] md:h-[418px] lg:h-[583px]"
+            />
+          </section>
+        </Wrapper>
+      </div>
+      <div className="flex flex-col justify-end overflow-hidden relative z-10  md:-mt-80 -mt-72 h-80 md:h-[420px] lg:h-[696px]">
+        <Image
+          alt="field"
+          src="/images/compliance/hero_1.jpg"
+          layout="fill"
+          objectPosition="top"
+          className="object-cover"
+          draggable={false}
+        />
+      </div>
     </section>
   );
 };
-
-export default Hero;
+export default Video;
