--- marketing/src/containers/compliance/hero/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
