+++ marketing/src/containers/compliance/science-based-targets/constants.ts
@@ -1,0 +1,33 @@
+import NATURE_SVG from 'svgs/compliance/nature.svg?sprite';
+import GLOBE_SVG from 'svgs/compliance/globe.svg?sprite';
+import GOALS_SVG from 'svgs/compliance/goals.svg?sprite';
+import DOCUMENTS_SVG from 'svgs/compliance/documents.svg?sprite';
+import CHART_SVG from 'svgs/compliance/chart.svg?sprite';
+
+export const TARGETS = [
+  {
+    icon: NATURE_SVG,
+    title: 'Assess Impact on Nature: ',
+    text: 'Use LandGriffon’s methodology to appraise nature impacts against SBTN-aligned indicators.',
+  },
+  {
+    icon: GLOBE_SVG,
+    title: 'Prioritize High-Risk Areas: ',
+    text: 'Allocate resources based on environmental materiality screening.',
+  },
+  {
+    icon: GOALS_SVG,
+    title: 'Set Impactful Science-Based Targets: ',
+    text: 'Establish specific targets across freshwater, land, climate, and biodiversity.',
+  },
+  {
+    icon: DOCUMENTS_SVG,
+    title: 'Make Informed Interventions: ',
+    text: "Follow SBTN's Action Framework (AR3T) for positive impact.",
+  },
+  {
+    icon: CHART_SVG,
+    title: 'Track Progress:',
+    text: 'Report publicly on progress using LandGriffon’s streamlined reporting tools.',
+  },
+];
