--- marketing/src/containers/compliance/video/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './component';
